requirements.txt
annotated-types==0.6.0
filelock==3.14.0
fsspec==2024.5.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.3.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dotenv==1.0.1
sympy==1.12
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.66.4
typing_extensions==4.11.0
